requirements.txt
asgiref==3.5.2
Django==4.1.3
django-dotenv==1.4.2
django-restframework==0.0.1
djangorestframework==3.14.0
mysqlclient==2.1.1
gunicorn==20.1.0
urllib3==1.26.17
whitenoise==6.2.0
django-environ==0.9.0
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
